remove grid fallbacks, remove old dashing grid, add grid-reverse class
