--- src/pages/components/css-grid/guidelines.js
@@ -1,25 +1,0 @@
-import React from 'react'
-
-import Layout from '../../../components/layout'
-import SubNav from '../../../components/subnavigation'
-import AppContent from '../../../components/appcontent'
-import StillWorking from '../../../components/stillworking'
-
-const currentPageName = "CSS Grid";
-const currentPageNameLower = currentPageName.toLowerCase();
-
-export default () => (
-  <Layout>
-    <header className="subnav">
-      <h1>{currentPageName}</h1>
-      <SubNav pageName={currentPageNameLower}/>
-    </header>
-    <AppContent>
-      <div className="row">
-        <div className="column column--full">
-          <StillWorking />
-        </div>
-      </div>
-    </AppContent>
-  </Layout>
-)
