--- src/paragon-framework/grid/_grid.scss
@@ -1,4 +1,0 @@
-@import 'grid-mixins';
-@import 'grid-fallbacks';
-@import 'grid-utilities';
-@import 'grid-main';
