--- src/paragon-framework/grid/grid-fallbacks.scss
@@ -1,49 +1,0 @@
-/* -----------------------------
-Fallback Classes (Must be First)
-This file is required for support on IE11 or Safari 9
-
-Standard sizes are named, non-standard sizes are numbered.
-
--------------------------------- */
-
-$gridConfig: '1'             '8.3333%',
-	'sixth'         '16.6666%',
-	'fourth'        '25%',
-	'third'         '33.3333%',
-	'5'             '41.6666%',
-	'half'          '50%',
-	'7'             '58.3333%',
-	'two-thirds'    '66.6666%',
-	'three-fourths' '75%',
-	'five-sixths'   '83.3333%',
-	'11'            '91.6666%',
-	'full'          '100%';
-
-//Standard Grid Classes
-@each $size in $gridConfig {
-	.grid--#{nth($size, 1)} {
-
-		@extend %fallback-gap;
-		@include fallback-grid(#{nth($size, 2)});
-	}
-}
-
-//Desktop Only Grid Classes
-@each $size in $gridConfig {
-	.grid--#{nth($size, 1)}_desktop {
-
-		@extend %fallback-gap;
-		@include fallback-grid_desktop(#{nth($size, 2)});
-	}
-}
-
-.grid {
-	> * {
-		float: left;
-		width: 100%;//By default, content will span all 12 columns
-	}
-}
-
-.grid-padding {
-	@extend %fallback-padding;
-}
