--- src/paragon-framework/grid/grid-main.scss
@@ -1,71 +1,0 @@
-/* -----------------------------
-Grid Classes
-
-Standard sizes are named, non-standard sizes are numbered.
--------------------------------- */
-$grid-gutter: 1rem !default;
-.grid { .card { margin: 0; } } //required to span cards 100%
-
-$gridConfig: '1'             '1',
-	'sixth'         '2',
-	'fourth'        '3',
-	'third'         '4',
-	'5'             '5',
-	'half'          '6',
-	'7'             '7',
-	'two-thirds'    '8',
-	'three-fourths' '9',
-	'five-sixths'   '10',
-	'11'            '11',
-	'full'          '12';
-
-.grid-wrapper {
-	margin: auto;
-	max-width: $global-content-width;
-	width: 100%;
-}
-
-.grid-layout,
-.grid-padding { padding: $grid-gutter; }
-
-.grid {
-	display: block;
-
-	grid-template-columns: repeat(12, 1fr);
-	overflow: auto;
-
-	* {
-		@supports (display: grid) {
-			@include grid-column(12);
-
-			width: auto;
-		}
-	}
-
-	//Make grid full width and centered on tablet
-	&.grid--tablet_full {
-		@media #{$tablet-and-below} {
-			margin: 0 auto;
-			max-width: 45rem;
-		}
-	}
-
-	@supports (display: grid) {
-		display: grid;
-		grid-gap: 0;
-
-		@media #{$tablet} { grid-gap: $grid-gutter; }
-	}
-
-	@each $size in $gridConfig {
-		.grid--#{nth($size, 1)} {
-			@include grid-column(#{nth($size, 2)});
-		}
-	}
-
-	@each $size in $gridConfig {
-		.grid--#{nth($size, 1)}_desktop {
-			@include grid-column_desktop(#{nth($size, 2)});
-		}
-	}
-}
