--- src/paragon-framework/grid/grid-mixins.scss
@@ -1,62 +1,0 @@
-@mixin fallback-grid($percent) {
-	@media #{$tablet} {
-		float: left; //SAFARI 9
-		width: calc(#{$percent} - #{$grid-gutter});
-	}
-}
-
-@mixin fallback-grid_desktop($percent) {
-	@media #{$desktop} {
-		float: left;
-		width: calc(#{$percent} - #{$grid-gutter});//SAFARI 9
-	}
-}
-
-@mixin grid-column($span) {
-	grid-column: 1 / -1;
-
-	@media #{$tablet} {
-		grid-column: span $span;
-
-		@supports (display: grid) { width: auto; }
-	}
-}
-
-@mixin grid-column_desktop($span) {
-	@media #{$desktop} {
-		grid-column: span $span;
-
-		@supports (display: grid) { width: auto; }
-	}
-}
-
-%fallback-gap {
-	margin: $grid-gutter 0;
-
-	&:first-of-type,
-	&:last-of-type { margin-top: 0; }
-
-	@media #{$tablet} {
-		margin: calc(#{$grid-gutter} / 2);
-
-		&:first-of-type,
-		&:last-of-type { margin: calc(#{$grid-gutter} / 2); }
-	}
-
-	@supports (display: grid) {
-		margin: 0;
-
-		&:first-of-type,
-		&:last-of-type { margin: 0; }
-	}
-}
-
-%fallback-padding {
-	padding: $grid-gutter 0;
-
-	@media #{$tablet} {
-		padding: calc(#{$grid-gutter} / 2);
-	}
-
-	@supports (display: grid) { padding: $grid-gutter; }
-}
