+++ src/paragon-framework/modules/_modules-components.scss
@@ -1,6 +1,5 @@
 @import 'typography/typography';
 @import 'grid/grid';
-// @import 'css-grid/css-grid'; // Replaced by component "grid"
 @import 'forms/forms';
 @import 'forms/themed-elements';
 @import 'forms/card-controls';
