+++ src/paragon-framework/modules/_modules.scss
@@ -1,11 +1,11 @@
 @import 'typography/typography';
 @import 'grid/grid';
-@import 'css-grid/css-grid'; // Replaced by component "grid"
+@import 'css-grid/css-grid';
 @import 'forms/forms';
 @import 'forms/themed-elements';
 @import 'forms/card-controls';
-@import 'actions/actions'; // Replaced by component "actions"
-@import 'actions/button-group'; // Replaced by component "actions"
+@import 'actions/actions';
+@import 'actions/button-group';
 @import 'avatar/avatar';
 @import 'pagination/pagination';
 @import 'tooltip/tooltip';
@@ -15,7 +15,7 @@
 @import 'table/table';
 @import 'navigation/navigation';
 @import 'navigation/mobile-menu';
-@import 'overpanel/overpanel';//dependent on header being first
+@import 'overpanel/overpanel'; //dependent on header being first
 @import 'footer/footer';
 @import 'icons/icons';
 @import 'spinner/spinner';
