--- src/paragon-framework/modules/css-grid/_css-grid.scss
@@ -1,4 +1,0 @@
-@import 'css-grid-mixins';
-@import 'css-grid-fallbacks';
-@import 'css-grid-utilities';
-@import 'css-grid-main';
