+++ src/paragon-framework/modules/grid/_grid.scss
@@ -1,0 +1,2 @@
+@import 'grid-mixins';
+@import 'grid-main';
