+++ src/paragon-framework/modules/grid/grid-main.scss
@@ -3,6 +3,9 @@ Grid Classes
 
 Standard sizes are named, non-standard sizes are numbered.
 -------------------------------- */
+$grid-gutter: 1rem !default;
+.grid { .card { margin: 0; } } //required to span cards 100%
+
 $gridConfig: '1'             '1',
 	'sixth'         '2',
 	'fourth'        '3',
@@ -65,6 +68,23 @@ $gridConfig: '1'             '1',
 			@include grid-column_desktop(#{nth($size, 2)});
 		}
 	}
-	
-	.card { margin: 0; }//remove margin for cards within grid
+}
+
+/* -----------------------------
+Utilities
+-------------------------------- */
+
+//Verticle Alignment
+.grid-align-center { align-items: center; }
+.grid-align-top { align-items: start; }
+.grid-align-bottom { align-items: end; }
+
+//Horizontal Alignment
+.grid-justify-center { justify-items: center; }
+.grid-justify-left { justify-items: start; }
+.grid-justify-right { justify-items: end; }
+
+//Direction (Reverse)
+.grid-reverse {
+  direction: rtl;
 }
