+++ src/paragon-framework/modules/grid/grid-mixins.scss
@@ -1,0 +1,17 @@
+@mixin grid-column($span) {
+	grid-column: 1 / -1;
+
+	@media #{$tablet} {
+		grid-column: span $span;
+
+		@supports (display: grid) { width: auto; }
+	}
+}
+
+@mixin grid-column_desktop($span) {
+	@media #{$desktop} {
+		grid-column: span $span;
+
+		@supports (display: grid) { width: auto; }
+	}
+}
